feat(cfn-include): preserve properties of resources that are not in the current CFN schema (#11822)
Currently, if cloudformation-include encounters a property in one of the resources the CDK has an L1 for
that is not in the version of the CloudFormation resource schema the module uses,
the property will be silently dropped. Of course, that is not ideal.

Make changes to our L1 code generation to preserve all properties that are not in the CFN schema.

Fixes #9717

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*