fix(core): floating list tokens synthesize to template
Encoded token lists are represented as `["#{Token[1234]}"]` (as opposed
to encoded strings, which are represented as `"${Token[1234]}"`, notice
the different sigil).

If a string that looks like `"#{Token[...]}"` ever appears outside
of a list context, that's an error. Someone got there by doing:

```
list[0]
```

Which we cannot allow, because things list `list[n]` are not detectable
by us and so cannot be adequately converted to CloudFormation syntax.

They are supposed to use `Fn.select(n, list)` instead.

We did not use to test for this, but of course people *will* do this
and then get confused about the results, so add an explicit test
and error message.

Closes #11750