chore(ecs-service-extensions): Adding defined return types to the ServiceExtension abstract class (#11916)
The ServiceExtension abstract class was missing a TypeScript definition for the return type of the hook. This worked through TypeScript type inference, but did not result in a helpful tooltip guide when attempting to implement your own ServiceExtension. This change should result in proper warnings and tooltips in IDE's when implementing a ServiceExtension

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*