feat(ecs): introduce a new Image type, TagParameterContainerImage, to be used in CodePipeline (#11795)
While CDK Pipelines is the idiomatic way of deploying ECS applications in CDK,
it does not handle the case where the application's source code is kept in a separate source code repository from the CDK infrastructure code.
This adds a new class to the ECS module,
`TagParameterContainerImage`, that allows deploying a service managed that way through CodePipeline.

Related to #1237
Related to #7746

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*