chore: use "canned metrics" obtained from CloudWatch Explorer (#11717)
CloudWatch Explorer is a new service who have been collecting a
set of metrics emitted by various services.

Integrate this model and use to generate convenience methods that are
aware of per-metric dimensions and suggested default statistics.

We switch over to these metrics to a get second source of
confirmation that we have the dimensions correct.

We can't do complete code generation off of this model yet. Adds a README document that explains
why.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*