chore(apigatewayv2): rename 'domainName' to 'name' in the DomainName construct
C# does not allow a property to have the same name as its class.
Jsii works around this by renaming the construct to `_DomainName` which
is ugly.

Instead, change the property name from `domainName` to `name` and avoid
this conflict entirely.

Further, re-categorize these class of defects as an error instead of a
warning.

BREAKING CHANGE: `domainName` property under `DomainName` has been
renamed to `name`.