fix(dynamodb): allow global replicas with Provisioned billing mode
We previously had validation inside the Table construct that would prevent creating global replicas
when billing mode was `PROVISIONED`.
However, it turns out that was too restrictive,
and it's actually possible to use the `PROVISIONED` mode,
but only if the Table has write auto-scaling enabled with a scaling policy.

Fixes #11346