fix(eks): failure to deploy cluster since aws-auth configmap exists (#12068)
The change in #12053 introduced a regression which causes failures in creating new clusters. Since we changed the KubernetesManifest resource to use `kubectl create` in CREATE operations, the attempt to create the `aws-auth` config map is failing because this config map is already created by the cluster.

This change adds an `override` to `KubernetesManifest` which will cause CREATE to be performed using `apply` instead, which practically allows overriding/adopting existing K8s resources.


----



*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*