refactor(aws-cdk): change the structure of the end-to-end tests
This PR changes the shape of the end-to-end
(previously called "integration")
tests that are present in the aws-cdk package.

The crux of the change is to introduce a new private NPM package that contains all of the test logic.
During local development,
this package will be managed by Yarn,
as all of the other CDK packages.
For our canaries,
it will be shipped as part of the aws-cdk package to NPM,
where our canaries can pick it up from.

The main idea is to encapsulate all of the details of how to run the integration tests in that package.
Instead of running some bash scripts inside of it directly,
the pipeline build and the canaries are now supposed to use the standard NPM commands to run these tests:

```
# unzip the files from the aws-cdk package...
cd e2e-tests
yarn install --production=true
yarn test-e2e-against-dist
```

This simplifies the setup of the tests quite a bit -
for example, we don't have to worry about installing ad-hoc dependencies,
like [we do](https://github.com/aws/aws-cdk/blob/2cd233a94fc2f3cb06211157738e59e8c7ee85e5/packages/aws-cdk/test/integ/common/jest-test.bash#L8)
 [now](https://github.com/aws/aws-cdk/blob/2cd233a94fc2f3cb06211157738e59e8c7ee85e5/packages/aws-cdk/test/integ/run-against-dist.bash#L48).
 It also allows a cleaner separation between the unit and end-to-end tests in the aws-cdk package.

 I've also taken this opportunity to change the names of the tests slightly,
 and divide them into separate directories,
 based on their type.