feat(ecs-patterns): Add DeploymentController option to Fargate services (#10452)
This change adds the option to set the `DeploymentController` on `ApplicationLoadBalanced`,
`NetworkLoadBalanced` and `QueueProcessing` ECS Services (both EC2 and Fargate). `MultipleTargetGroups` services are not updated because this option might not be applicable for them - one service might be using the default deployment
controller while others might use `CODE_DEPLOY` or `EXTERNAL`.

By default if this options is not passed it will use the default one from the respective Service construct  which
is `ECS`'s rolling updates.

Related to aws/containers-roadmap#130

Closes #10971.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*