feat(eks): bundle kubectl, helm and awscli instead of SAR app
Bundle the AWS Lambda layer zip bundle as part of the EKS module and upload as an asset to the destination account in order to remove the dependency on the [aws-lambda-layer-kubectl](https://github.com/aws-samples/aws-lambda-layer-kubectl) SAR app.

The dependency on the SAR app introduces an operational and maintenance risk, and increases deploy time due to an additional nested stack introduced by SAR.

This also ensures that the EKS module can be deployed to any environment, regardless of whether the SAR app is avaialble in that location.

This change increases the module size by ~40MiB.

Fixes #11874