chore: compat script will now also install peer dependencies
In cdkv2, the package `aws-cdk-lib` has changed its dependency model to
move to use peerDependencies + devDependencies, instead of regular
dependencies.

As part of this, the `constructs` package is now a peer dependency.
jsii-diff requires this package to also be installed to load the jsii
assembly.

Change the check-api-compatibility.sh to also install all declared peer
dependencies before running jsii-diff.