chore: compat script will now also install peer dependencies (#12142)
In cdkv2, the package `aws-cdk-lib` has changed its dependency model to
move to use peerDependencies + devDependencies, instead of regular
dependencies.

As part of this, the `constructs` package is now a peer dependency.
jsii-diff requires this package to also be installed to load the jsii
assembly.

Change the check-api-compatibility.sh to use npm7 which installs all
declared peer dependencies as part of `npm install`.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*