fix(core): capturing stack traces still takes a long time (#12180)
Stack trace captured was intended to be disabled by default in #11170.

However, due to a logic error that didn't actually happen. This change
*actually* disables stack trace capture.

The intent of the original change was to capture *Token* stack traces if
and only if `CDK_DEBUG=true` was set. *Metadata* and *Construct* stack
traces are managed in the `constructs` library, no longer in CDK.

BREAKING CHANGE: Creation stack traces for `Lazy` values are no longer
captured by default in order to speed up tests. Run with
`CDK_DEBUG=true` (or `cdk --debug`) to capture stack traces.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*