fix(dynamodb): allow global replicas with Provisioned billing mode (#12159)
We previously had validation inside the Table construct that would prevent creating global replicas
when billing mode was `PROVISIONED`.
However, it turns out that was too restrictive,
and it's actually possible to use the `PROVISIONED` mode,
but only if the Table has write auto-scaling enabled with a scaling policy.

Fixes #11346

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*