chore(s3-deployment): restore lambda tests
The change in #12129 accidentally deleted the unit tests for the s3-deployment lambda function. This change restores them.

In order to allow tests to stub the `aws` CLI, I've added an environment variable TEST_AWSCLI_PATH. If it is defined, it will be determine the path to the CLI. Otherwise, `/opt/awscli/aws` will be used.