feat(cdk-assets): add external asset support
In the event that assets are not actually available at synthesis
time, we still want to support JIT (just-in-time) asset
generation via external tooling. This would, for instance, enable
a third party tool to fetch additional resources prior to
bundling/building and subsequent uploading.

This adds a new interface for both File and Docker asset types
that allows users to specify an executable. The executable,
depending on the asset type, must then reply with a specific key
on stdout, which will then get picked up and used by CDK Assets.

This also updates the default stack synthesizer to support adding
external sources directly. This is technically a breaking change
for anyone who currently extends the class.