fix(eks): updated the eks launch template support topic in the docs
fixes #12256

Clarified the usage of custom user data with default or custom AMI by adding another example and pointing out the differences.

Also changed the launch template reference in the examples to use the latest version rather than the default. If you don't know the difference it would be unexpected for an user to change the user data without having it applied to the nodes after deploy.

Make the sentence about defining instance type either in launch template or node group more understandable.