fix(aws-cdk): remove 'useWordWrap' to prevent CLI error
fixes: #12306

The table dependency recently updated their validation
to not allow additional properties.
The CloudFormation diff however, seems to be using an unknown property `useWordWrap`.
This results in the CLI terminating prematurely and printing an error in the table dependency.

As `useWordWrap` is nowhere to be found in the table package,
this commit simply removes it.