chore(cli): IAM differences table printing is broken (#12330)
We were passing (by mistake) an unrecognized property to the `table` module.

This didn't use to be a problem, but the `table` module has suddenly
started throwing an "invalid config" error. The cause is a dependency
update that happened today.

* In #12324, the version of `table` was updated from `6.0.4` to `6.0.7`.
* In [6.0.5](https://github.com/gajus/table/releases/tag/v6.0.5),
  `table` changes the way it does JSON validation of the input
  configuration ob ject.

Didn't dive into all of the changes, but it stands to reason the
validator has become more strict and where it *used* to allow additional
properties (probably by accident) it no longer does so.

Undo the passing of the unused property.

This change is a **chore** instead of a **fix** because the change
wasn't released yet.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*