fix(ec2): interface endpoint AZ lookup does not guard against broken situations (#12033)
`lookupSupportedAzs` can create a VPC endpoint object with no subnets, or fewer subnets than expected, if AZs are not resolved (i.e. they are Tokens). This can create deployment-time failures if all the VPC subnets are filtered out.

This is not a problem with stock CDK, this is a problem with third-party CDK libraries that do not provide concrete AZs (us-east-1a, us-east-1b) for subnets, even though the account and region are specified.

A less common problem is that `lookupSupportedAzs` can return no subnets if there is no overlap between the user's VPC AZs and the AZs of a service they're connecting to. This also results in a deployment-time failure.

This commit adds error checking to make sure that all subnet AZs are concrete values when using `lookupSupportedAzs`, and also throws an error if the user inadvertently attempts to create a VPC endpoint with no subnets. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*