chore(lambda): throw a relevant error message when asset is used across stacks (#12420)
Although the constructor of the AssetCode doesn't imply it's real
dependencies, an AssetCode currently can only be associated to consumers
within a single stack.

Attempting to use them across stacks results in the conspicuous error -
"Cannot reference across apps. Consuming and producing stacks must be
defined within the same CDK app."

The correct solution here would have been to model this such that the
stack dependency is made explicit on the customer facing API, perhaps by
taking a 'scope' parameter.

For the moment, throw a more relevant error message so that it's obvious to
the user on what is wrong.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*