chore(scripts): compat check will compare against last published version
The script currently does not support publishing CDKv2.

Update the script so as to read the baseline version by resolving the
version from the repo's current state (from `resolve-version.js`).

If the resolved version is not present in NPM, usually because the
version is not *yet* published, fallback to an environment variable
NPM_DISTTAG that specifies the dist tag to use.