chore(scripts): compatibility check filters on package version
The script will fail in a PR build for the following scenario -
The CDK version is at 1.x.0 and an import of the CloudFormation spec
has introduced a new CDK package.
Subsequently, a PR is created against master (still at 1.x.0).
Subsequent releases of the CDK publishes this package into npm, and
the version has moved to 1.y.0 (where y > x).
However, the original PR is not yet updated with master, and continues
to be updated.

In such a scenario, this script will fail sine the new CDK package is
present in npm but not present at the resolved version (1.x.0).

Update the filtering process to also include the resolved version.