chore(eks): updated the eks launch template support topic in the docs (#12272)
Closes https://github.com/aws/aws-cdk/issues/12256

Clarified the usage of custom user data with default or custom AMI by adding another example and pointing out the differences.

Changed the launch template reference in the examples to use the latest version rather than the default. If you don't know the difference it would be unexpected for the user to change the user data without having it applied to the nodes after deploy.

Removed unnecessary wrapping of instance type string with typed version and calling toString.

Make the sentence about defining instance type either in launch template or node group more understandable.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*