fix(iam): Groups are erroneously accepted as the Principal of a policy (#11479) (#12549)
closes #11479
----
Hello, picking this up as my first issue to learn about how CDK works:

[AWS mentions](https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_elements_principal.html) that iam groups can't be principals in policies. If a developer doesn't know this, they get an error during deployment. This change will throw an error to help them catch the issue during synthesis time.

Note that this doesn't handle the code path where an arn string is specified as the principal, but the method already documents that a group would be invalid.

Let me know if there's a cleaner way to fix this.