feat: reset all current feature flags for CDKv2
For all feature flags introduced in CDKv1, the default behaviour in
CDKv2 is the one where the flag is turned on.

Except for `STACK_RELATIVE_EXPORTS_CONTEXT` and
`NEW_STYLE_STACK_SYNTHESIS_CONTEXT`, the rest of the feature flags are
inconfigurable in CDKv2, and cannot be reverted back to the old
behaviour.

`STACK_RELATIVE_EXPORTS_CONTEXT`: Since CloudFormation prevents changes
and deletion of stack exports that are consumed, this feature flag is
retained. If not, customer will have to re-create their stacks.

`NEW_STYLE_STACK_SYNTHESIS_CONTEXT`: This feature is not widely
available yet. Further, customers will need to re-bootstrap their
accounts, which in some cases may not be possible.