chore: make the 'constructs' version used in the e2e-tests configurable
The monocdk packages in the V2 branch were
[recently switched](https://github.com/aws/aws-cdk/pull/12054)
to using constructs in version 10.0.0-pre.5.
However, our e2e-tests always
[install the latest constructs](https://github.com/aws/aws-cdk/blob/12868a2ac9c7016e47633045fb0c00be7eeb5092/packages/aws-cdk/test/integ/helpers/cdk.ts#L124),
which is 3.2.108 currently, as pre-release versions are not considered when installing the latest
(and if they were, that would break the tests on the 'master' branch anyway).

Make the version of 'constructs' used configurable through an environment variable.
We will then need to update the tests in the v2 branch to set that to 10.0.0-pre.5.