chore: compat check does not correctly resolve baseline version in release PRs
In release PRs, the resolved version will be the next minor or patch
version to be released. This will occur on bump PRs on the 'release' and
'v2-main' branches.

The existing logic for such cases was to fall back to `NPM_DISTTAG`
environment variable. However, the value needs to be different in the
'release' branch and the 'v2-main' branches.

Configuring this in our current infrastructure is not easy. Also, this
makes it one step harder to configure new branches and pipelines, as
the set up would need to decide what the value of `NPM_DISTTAG` should
be.

Instead, taking a simpler approach where all branches will have two
environment variables configured - `NPM_DISTTAG_V1` and `NPM_DISTTAG_V2`.

`check-api-compatibility.sh` will determine which of the two to use
based on the major version.