chore: install integ tests dependencies using npm7
This has the advantage that the right `peerDependencies` versions are
automatically installed without additional effort, which will benefit
the `v2` test.

Remove the `npm` wrapper: instead we'll write a `package.json` that
either contains package names or paths to directories in the monorepo.
Same effect, but less dependent on the exact `npm` command that gets
run.

We then switch to `npm7` to install the files in `package.json`.