chore: install integ tests dependencies using npm7 (#12626)
This has the advantage that the right `peerDependencies` versions are
automatically installed without additional effort, which will benefit
the `v2` test.

Remove the `npm` wrapper: instead we'll write a `package.json` that
either contains package names or paths to directories in the monorepo.
Same effect, but less dependent on the exact `npm` command that gets
run.

We then switch to `npm7` to install the files in `package.json`.

If all goes well, supersedes #12605 without having additional configuration.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*