chore: remove strong-named key and signing for .NET
The benefits of strong-name signing assemblies are marginal at best, and it
turns out we actually weren't strong-name signing our assemblies anyway, making
all of this extra infrastructure unnecessary.

This removes the strong-name key (SNK) prebuild script, updates the buildspec
to remove the reference to it, and removes the linter rule that enforced the
strong-named assembly properties. Prior to removal, the rule was reversed and
used to update the package.json files.

This change will be accompanied by a change to our pipeline definition to
remove the references on that side.