fix(apigateway): cannot remove first api key from usage plan (#12505)
This is caused by a [commit introduced] back in Nov 2019.
The original change tried to be overly ambitious in not triggering a
resource replacement on customers who already had a single key usage
plan configured.

This results in stack update failures for a usage plan with multiple
keys, when the **first** key is removed. Removal of the first key
re-adjusts the logical ids of all the keys in a way that makes the
update look like the first key is changing and a subsequent key is being
removed.

The fix is to simply not special case the first key.

Resource replacement of Usage Plan Key resource should not create any
impact on running applications.

fixes #11876

[commit introduced]: https://github.com/aws/aws-cdk/commit/142bd0e2fe33de239fc34f72d7f74aab81457607

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*